verification/integration/requirements.txt
pytest-asyncio==0.21.0
-r ../../zephyr/scripts/requirements.txt
